libcamera: sc8830: Squashed fixing compilation errors
Picked commits:
https://github.com/remilia15/android_hardware_sprd/commit/6f56e9da2e3c52889e8f1b0adb62664e61e0ce40
https://github.com/remilia15/android_hardware_sprd/commit/4987669e9264d50243b6f125a71b46b09bbfff39
https://github.com/remilia15/android_hardware_sprd/commit/be08a46782e9e50f8fb920c91f64ecac44d82d44
https://github.com/remilia15/android_hardware_sprd/commit/d460776593fc0e8bc508982862b69e8cb5e423a4
https://github.com/remilia15/android_hardware_sprd/commit/fcc93bfa6d05f633ef0120ce7b9cca966e2f5fb1
https://github.com/remilia15/android_hardware_sprd/commit/ca53db0a808408fb2fde3deca2506f129f5f0332

+++ libcamera/sc8830/isp2.0/param_manager/isp_com_alg.c
@@ -540,15 +540,15 @@ int32_t isp_scaling_lsc_gain(uint16_t *dst, uint16_t *src, struct isp_size *dst_
 			src_x >>= 16;
 
 			if (src_x < dst_size->w - 1) {
-				src_l[0] = src_line[0] +src_x;
-				src_l[1] = src_line[0] + src_x + 1;
-				src_l[2] = src_line[1] +src_x;
-				src_l[3] = src_line[1] +src_x + 1;
+				src_l[0] = (uint32_t)src_line[0] +src_x;
+				src_l[1] = (uint32_t)src_line[0] + src_x + 1;
+				src_l[2] = (uint32_t)src_line[1] +src_x;
+				src_l[3] = (uint32_t)src_line[1] +src_x + 1;
 			} else {
 				src_x = dst_size->w - 1;
-				src_l[0] =src_line[0] + src_x;
+				src_l[0] = (uint32_t)src_line[0] + src_x;
 				src_l[1] = src_l[0];
-				src_l[2] =src_line[1] + src_x;
+				src_l[2] = (uint32_t)src_line[1] + src_x;
 				src_l[3] = src_l[3];
 			}
